fix(kysely): register the schema check inside the adapter factory
