+++ packages/kysely-adapter/src/index.ts
@@ -1,5 +1,11 @@
 export * from "./dialect";
 export * from "./kysely-adapter";
+export {
+	getMssqlSchema,
+	getPostgresSchema,
+	toIntrospectedTables,
+	toPhysicalSchema,
+} from "./schema-check";
 export * from "./types";
 // Don't export node:sqlite by default, as it is not production ready.
 // export * from "./node-sqlite";
